could an error handler avoid cascading failures?
This is just a theory, and this code is untested, but
it should be enough to evaluate for now.

I suspect that when one of the services dies, the killed
stream turns into an unhandled exception, and kills the
proxy.

This then might be responsible for random services jumping
in and out of existence, and why the router state sometimes
seems out of sync.

There might be other knockon effects here too.